tests/wpt/web-platform-tests/css/css-logical/parsing/margin-block-inline-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Logical Properties and Values: parsing margin-block and margin-inline with invalid values</title>
<link rel="help" href="https://drafts.csswg.org/css-logical/#propdef-margin-block">
<meta name="assert" content="margin-block, margin-inline support only the grammar '<'margin-top'>{1,2}'.">
<meta name="assert" content="margin-block, margin-inline longhands support only the grammar '<'margin-top'>'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("margin-block-start", "none");
test_invalid_value("margin-block-end", "10");
test_invalid_value("margin-inline-start", "20% calc(10px - 0.5em)");
test_invalid_value("margin-inline-end", "10px, auto");

test_invalid_value("margin-block", "none");
test_invalid_value("margin-block", "20%, calc(10px - 0.5em)");
test_invalid_value("margin-inline", "10px auto 20px");
</script>
</body>
</html>
